// apps/admin/src/types/DataTable.ts
export interface SearchParams {
    [key: string]: string | string[] | undefined | null
  }
  
  export type Option = {
    label: string
    value: string
    icon?: React.ComponentType<{ className?: string }>
  }
  
  export interface DataTableFilterOption<TData> {
    id?: string
    label: string
    value: keyof TData | string
    items: Option[]
    isMulti?: boolean
  }
  
  export interface DataTableSearchableColumn<TData> {
    id: keyof TData
    title: string
    isDate?: boolean
  }
  
  export interface DataTableFilterableColumn<TData>
    extends DataTableSearchableColumn<TData> {
    options: Option[]
  }
  